refactor(integrations): extract _shared/{sink,session} from openai (Python half)
Lifts TraceSink protocol + FileSink + autocontext_session contextvar from
the openai-specific integration into a new autocontext.integrations._shared
package so future integrations (anthropic, langchain, ...) consume them
directly. A2-II-b public API preserved via re-exports at the openai module
top level. Internal imports within openai package redirected to thin
re-export stubs in _sink.py / _session.py. Test files moved to
tests/integrations/_shared/. No behavior change; openai integration tests
all passing.
